+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0 - 2025-03-30
+
+- Changed `@cypress/unique-selector` to `@guoyunhe/unique-selector` to support ESM
+- Added `DOMLint.print()` method
+
 ## 0.1.0 - 2025-03-30
 
 - Added `DOMLint` class